complete hibernate task

+++ src/main/java/com/epam/model/Training.java
@@ -1,9 +1,9 @@
 package com.epam.model;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import jakarta.persistence.*;
 import java.time.LocalDate;
 
 @Getter
@@ -35,5 +35,5 @@ public class Training {
     private LocalDate trainingDate;
 
     @Column(name = "training_duration", nullable = false)
-    private int trainingDuration;
+    private Double trainingDuration;
 }
